Make UserAccess#grant_access more efficient
- bulk import
- check from memory instead of DB

+++ app/models/user_access.rb
@@ -62,23 +62,25 @@ class UserAccess < Struct.new(:user)
   end
 
   def permitted_access_levels
-    return LEVELS.keys if user.power_user?
+    return LEVELS.keys if bypass?
 
     LEVELS[user.access_level.to_sym][:grant_access]
   end
 
   def grant_access(new_user, selected_facility_ids)
     return if selected_facility_ids.blank?
+    return if bypass?
+
     raise NotAuthorizedError unless permitted_access_levels.include?(new_user.access_level.to_sym)
 
-    resources = prepare_access_resources(selected_facility_ids)
+    resources = prepare_grantable_resources(selected_facility_ids)
     # if the user couldn't prepare resources for new_user means they shouldn't have had access to this operation at all
     raise NotAuthorizedError if resources.empty?
 
     # recreate accesses from scratch to handle deletes/edits/updates seamlessly
     User.transaction do
       new_user.accesses.delete_all
-      new_user.accesses.create!(resources)
+      new_user.accesses.import!(resources)
     end
   end
 
@@ -89,14 +91,14 @@ class UserAccess < Struct.new(:user)
   private
 
   def can_access_record?(resources, record)
-    return true if user.power_user?
+    return true if bypass?
     return resources.find_by_id(record).present? if record
 
     resources.exists?
   end
 
   def resources_for(resource_model, action)
-    return resource_model.all if user.power_user?
+    return resource_model.all if bypass?
     return resource_model.none unless ACTION_TO_LEVEL.fetch(action).include?(user.access_level.to_sym)
 
     resource_ids =
@@ -108,28 +110,43 @@ class UserAccess < Struct.new(:user)
     resource_model.where(id: resource_ids)
   end
 
-  def prepare_access_resources(selected_facility_ids)
-    selected_facilities = Facility.where(id: selected_facility_ids)
+  #
+  # Compare the new user's selected facilities with the currently accessible facilities of the current user
+  # and see if we can promote the new user's access if necessary
+  def prepare_grantable_resources(selected_facility_ids)
+    selected_facilities = Facility.where(id: selected_facility_ids).includes(facility_group: :organization)
     resources = []
 
+    accessible_facilities_in_org = accessible_facilities(:manage).group_by(&:organization)
     selected_facilities.group_by(&:organization).each do |org, selected_facilities_in_org|
-      if can?(:manage, :organization, org) && org.facilities == selected_facilities_in_org
-        resources << {resource: org}
+      if can?(:manage, :organization, org) &&
+        (accessible_facilities_in_org[org].to_set == selected_facilities_in_org.to_set)
+
+        resources << {resource_type: Organization.name, resource_id: org.id}
         selected_facilities -= selected_facilities_in_org
       end
     end
 
+    accessible_facilities_in_fg = accessible_facilities(:manage).group_by(&:facility_group)
     selected_facilities.group_by(&:facility_group).each do |fg, selected_facilities_in_fg|
-      if can?(:manage, :facility_group, fg) && fg.facilities == selected_facilities_in_fg
-        resources << {resource: fg}
+      if can?(:manage, :facility_group, fg) &&
+        (accessible_facilities_in_fg[fg].to_set == selected_facilities_in_fg.to_set)
+
+        resources << {resource_type: FacilityGroup.name, resource_id: fg.id}
         selected_facilities -= selected_facilities_in_fg
       end
     end
 
     selected_facilities.each do |f|
-      resources << {resource: f} if can?(:manage, :facility, f)
+      if can?(:manage, :facility, f)
+        resources << {resource_type: Facility.name, resource_id: f.id}
+      end
     end
 
     resources.flatten
+  end
+
+  def bypass?
+    user.power_user?
   end
 end
